requirements_p3d.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.9.2=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.5.0=py37h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
autopage=0.5.1=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blosc=1.21.0=h8c45485_0
bottleneck=1.3.5=py37h7deecbd_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
cfgv=3.3.1=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cliff=3.10.1=pypi_0
cmaes=0.9.1=pypi_0
cmd2=2.4.3=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
colorlog=6.7.0=pypi_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.28=py37h295c915_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distlib=0.3.6=pypi_0
einops=0.6.0=pypi_0
entrypoints=0.4=py37h06a4308_0
exceptiongroup=1.1.0=pypi_0
expat=2.4.4=h295c915_0
filelock=3.9.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.38.0=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20210915=py37
gensim=4.2.0=py37h6a678d5_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=2.0.2=pypi_0
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.8.0=pypi_0
hdf5=1.12.1=h69dfa17_1
hydra-colorlog=1.2.0=pypi_0
hydra-core=1.2.0=pypi_0
hydra-optuna-sweeper=1.2.0=pypi_0
icu=58.2=he6710b0_3
identify=2.5.17=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.25.0=pypi_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=py37
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py37h06a4308_0
jupyter=1.0.0=py37h06a4308_8
jupyter_client=7.2.2=py37h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py37h06a4308_0
jupyter_server=1.18.1=py37h06a4308_0
jupyterlab=3.4.4=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py37h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=7.84.0=h91b91d3_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
lzo=2.10=h7b6447c_2
mako=1.2.4=pypi_0
markdown=3.4.1=pypi_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mdurl=0.1.2=pypi_0
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mock=4.0.3=pyhd3eb1b0_0
multidict=6.0.4=pypi_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.3.0=py37h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nodeenv=1.7.0=pypi_0
notebook=6.4.12=py37h06a4308_0
numexpr=2.8.3=py37h807cd23_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
nvidiacub=1.10.0=0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.7.0.68=pypi_0
openssl=1.1.1q=h7f8727e_0
optuna=2.10.1=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pbr=5.11.1=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=22.1.2=py37h06a4308_0
platformdirs=2.6.2=pypi_0
pluggy=1.0.0=pypi_0
portalocker=2.5.1=py37h89c1867_0
pre-commit=2.21.0=pypi_0
prettytable=3.6.0=pypi_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.20.3=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.1=pypi_0
pygments=2.14.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyrootutils=1.0.4=pypi_0
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
pytables=3.7.0=py37h9da3b7f_0
pytest=7.2.1=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.1=pypi_0
python-fastjsonschema=2.16.2=py37h06a4308_0
python_abi=3.7=2_cp37m
pytorch=1.7.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytorch-lightning=1.5.0=pypi_0
pytorch3d=0.6.2=py37_cu102_pyt171
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h540881e_4
pyzmq=23.2.0=py37h6a678d5_0
qt=5.9.7=h5867ecd_1
qtconsole=5.3.1=py37h06a4308_1
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rich=13.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.7.3=py37hc147768_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py37h06a4308_0
sh=1.14.3=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py37h06a4308_0
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlalchemy=2.0.1=pypi_0
sqlite=3.39.2=h5082296_0
stevedore=3.5.2=pypi_0
tabulate=0.8.10=pyhd8ed1ab_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pyhd8ed1ab_3
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchmetrics=0.10.3=pypi_0
torchvision=0.8.2=py37_cu102
tornado=6.1=py37h27cfd23_0
tqdm=4.64.0=py37h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
trimesh=3.13.5=pyh6c4a22f_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.11=py37h06a4308_0
virtualenv=20.17.1=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py37h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
